done without plate finding

+++ assign_3/get_circle_mask.m
@@ -1,5 +1,5 @@
 function [MASK] = get_circle_mask(sx, sy, r, cx, cy)
-    MASK = zeros(sx, sy);
+    MASK = uint8(zeros(sx, sy));
     for x = 1:sx
         for y = 1:sy
             if (x-cx)^2 + (y-cy)^2 < r^2
